requirements.txt
cached-property==1.5.2
certifi==2020.6.20
chainer==7.7.0
chainerrl==0.8.0
cloudpickle==1.6.0
cupy-cuda101==7.7.0
cycler==0.10.0
decorator==4.4.2
fastrlock==0.5
filelock==3.0.12
future==0.18.2
gym==0.17.3
imageio==2.9.0
kiwisolver==1.2.0
matplotlib==3.3.2
##mkl-fft==1.0.6
##mkl-random==1.0.1.1
#mkl-service==2.3.0
networkx==2.5
numpy==1.19.2
olefile==0.46
opencv-python==4.4.0.44
Pillow 
protobuf==3.13.0
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.3
#six==1.15.0
tifffile==2020.9.3
#torch==1.6.0
torchvision==0.7.0
typing-extensions==3.7.4.3
